+++ R/RECOM_BIN_AR.R
@@ -2,11 +2,12 @@
 
 .BIN_AR_param <- list(
   support = 0.1,
-  confidence = 0.3,
-  maxlen = 5,
-  measure = "confidence",
-  verbose = FALSE,
-  decreasing = TRUE
+  confidence = 0.8,
+  maxlen = 3,
+  sort_measure = "confidence",
+  sort_decreasing = TRUE,
+  apriori_control = list(),
+  verbose = FALSE
 )
 
 BIN_AR <- function(data, parameter = NULL) {
@@ -14,23 +15,28 @@ BIN_AR <- function(data, parameter = NULL) {
   ## parameters
   p <- getParameters(.BIN_AR_param, parameter)
 
+  p$apriori_control$verbose <- p$verbose
+
   ## binaryRatingMatrix stores arules transactions!
   data <- data@data
 
-  rule_base <- apriori(data,
+  ## new version of arules warns for maxlen constraint
+  rule_base <-  suppressWarnings(
+    apriori(data,
     parameter=list(support=p$support, confidence=p$confidence,
-      minlen=2, maxlen=p$maxlen), control=list(verbose=p$verbose))
+      minlen=2, maxlen=p$maxlen), control=p$apriori_control)
+    )
 
-  ## additional measures
-  if(p$measure == "cxs") quality(rule_base) <- cbind(quality(rule_base),
+  ## additional measures for sorting the rulebase
+  if(p$sort_measure == "cxs") quality(rule_base) <- cbind(quality(rule_base),
     cxs = quality(rule_base)$confidence * quality(rule_base)$support)
 
-  if(!p$measure %in% names(quality(rule_base))) quality(rule_base) <-
-    cbind(quality(rule_base), interestMeasure(rule_base, method = p$measure,
+  if(!p$sort_measure %in% names(quality(rule_base))) quality(rule_base) <-
+    cbind(quality(rule_base), interestMeasure(rule_base, method = p$sort_measure,
       transactions = data))
 
   ## sort rule_base
-  rule_base <- sort(rule_base, by = p$measure, decreasing=p$decreasing)
+  rule_base <- sort(rule_base, by = p$sort_measure, decreasing=p$sort_decreasing)
 
   if(p$verbose) cat("\nRule base size:", length(rule_base), "rules.\n")
 
@@ -55,13 +61,13 @@ BIN_AR <- function(data, parameter = NULL) {
     }
 
     n <- as.integer(n)
-    measure <- model$measure
+    sort_measure <- model$sort_measure
 
     reclist <- list()
     m <- is.subset(lhs(model$rule_base), newdata@data)
     for(i in 1:nrow(newdata)) {
       recom <- head(unique(unlist(
-        LIST(rhs(sort(model$rule_base[m[,i]], by=measure)),
+        LIST(rhs(sort(model$rule_base[m[,i]], by=sort_measure)),
           decode=FALSE))), n)
 
       reclist[[i]] <- if(!is.null(recom)) recom else integer(0)